Network 캐릭터의 경우 서버를 통해 HP 를 관리하도록 수정

+++ Project_Animagus/Server/Server/Player.h
@@ -22,7 +22,7 @@ public:
 
 public:
     uint16  playerID = 0;
-    int8    playerHP = 100;
+    int16    playerHP = 100;
     float   x, y, z;
     float   rotation;   // y축 회전각
